+++ components/__init__.py
@@ -1,0 +1,5 @@
+import os
+import sys
+
+project_root = os.path.dirname(os.path.realpath(__file__))
+sys.path.append(os.path.join(project_root, 'components'))
