+++ components/get_exif.py
