+++ components/parsers/exif/utils.py
